Fixed a warning when formatting a double variant to string. This might have been an error tho, as you cannot use SHVInt, SHVDouble etc in a ... operator, you can only use C style types.

+++ shivamodulelibs/dataengine/src/shvdatavariantimpl.cpp
@@ -405,7 +405,7 @@ void SHVDataVariantImpl::SetDouble(SHVDouble val)
 			break;
 		case SHVDataVariant::TypeString:
 			Data.stringVal = new SHVString();
-			Data.stringVal->Format(_T("%g"), val);
+			Data.stringVal->Format(_T("%g"), (double)val);
 			break;
 		}
 	}
